Breakthrough
Finally fixed the board reset problem. Took me HOURS and was because of
a single line of stupid code. Valid move checker and win checker both
work now. Board updates as it should now. See Randy.createNew() for an
example on how to pass setting from the Agent's JPanel to a newly
created Agent.