Game Loop Reworked and Simplified
Game Loop in Control is cleaner and works as expected. As per the
requirements, player X always makes the first move. player1turn() is now
a property of a state-space and not a manual boolean to set after each
turn/win. Mason's heuristic works but doesn't play defensively (to block
a win).